add bigtiff flag to output profile for exg and ndvi data products, skip sending mail when testing
